console

+++ src/main/java/io/github/shiruka/shiruka/ShirukaMain.java
@@ -34,8 +34,6 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import org.jetbrains.annotations.NotNull;
-import org.jline.reader.LineReaderBuilder;
-import org.jline.terminal.TerminalBuilder;
 
 /**
  * a Java main class to start the Shiru ka's server.
@@ -122,7 +120,8 @@ public final class ShirukaMain {
               // not equal the current fragment's version delete file and download the latest.
               Files.delete(fragmentPath);
               return true;
-            } else {
+            } else if (Files.exists(fragmentPath)) {
+              // If the fragment file exist
               // If the fragment should not download and load delete the file.
               Files.delete(fragmentPath);
             }
@@ -140,22 +139,7 @@ public final class ShirukaMain {
         }));
     Shiruka.initServer(server, fragmentManager);
     fragmentManager.loadFragments();
-    final var reader = LineReaderBuilder.builder()
-      .appName("Shiru ka")
-      .terminal(TerminalBuilder.builder()
-        .dumb(true)
-        .jansi(true)
-        .build())
-      .build();
-    while (true) {
-      final var line = reader.readLine("$ ");
-      if (line.isEmpty()) {
-        continue;
-      }
-      server.runCommand(line);
-      if (server.isInShutdownState()) {
-        return;
-      }
-    }
+    final var console = new ShirukaConsole(server);
+    console.start();
   }
 }
